Return error if no objects in deletion request
This would be imposible to reach normally, however if a user manually removes the disabled class from the delete button and uses it before selecting, this will be thrown

+++ raptorWeb/panel/views.py
@@ -648,9 +648,12 @@ class PanelDeleteView(View):
         changed_string: str = ''
         for model in deleting_objects:
             changed_string += f'{model}, '
+            
+        if changed_string == '':
+            messages.error(request, 'There were no objects to delete!')
+            return HttpResponse(status=400)
 
         deleting_objects.delete()
-
         PanelLogEntry.objects.create(
             changing_user=request.user,
             changed_model=str(f'{model_string} - {changed_string}'),
